requirements.txt
configparser==3.3.0.post2
pika==0.10.0
pkg-resources==0.0.0
pyserial==2.7
PyYAML==3.12
requests==2.7.0
serialenum==0.3
